AiXRayFilmDetectionSystem/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
SQLAlchemy>=2.0.36
PyJWT==2.8.0
Werkzeug==2.3.7
torch>=2.6.0
torchvision>=0.21.0
opencv-python>=4.8.1.78
Pillow>=10.1.0
numpy>=1.25.2
openai>=1.12.0
gunicorn>=21.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
